requirements-lint.txt
homeassistant-stubs
mypy
pre-commit
ruff
types-aiofiles
types-cffi
types-PyMySQL
types-pyRFC3339
types-python-dateutil
types-PyYAML
types-requests
types-setuptools
